refactor(budget): v1 API 라우트 제거 및 v2 정식 경로 승격
- /api/budget/route.ts (v1) 제거
- today/, monthly/ 정식 라우트 신규 생성
- runway/, settings/ v2 로직으로 교체
- v2/ 스테이징 디렉토리 제거
- 프론트엔드 fetch URL: /v2/* → 정식 경로로 업데이트
