+++ web/src/app/api/budget/monthly/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import { requireAuth } from '@/lib/auth';
-import { getTodayAllocation } from '@/features/budget/lib/facade';
+import { getMonthlyAllocation } from '@/features/budget/lib/facade';
 
 export const dynamic = 'force-dynamic';
 
@@ -9,10 +9,10 @@ export async function GET() {
   if (!userId) return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
 
   try {
-    const data = await getTodayAllocation(userId, new Date());
+    const data = await getMonthlyAllocation(userId, new Date());
     return NextResponse.json({ data });
   } catch (err) {
-    console.error('[Budget v2 API] today', err);
-    return NextResponse.json({ error: '오늘 예산 조회 실패' }, { status: 500 });
+    console.error('[Budget API] monthly', err);
+    return NextResponse.json({ error: '월 예산 조회 실패' }, { status: 500 });
   }
 }
