+++ web/src/features/budget/components/budget-settings-page.tsx
@@ -285,7 +285,7 @@ function TargetDateCard({
     if (!/^\d{4}-\d{2}$/.test(target)) return;
     setLoadingPreview(true);
     try {
-      const res = await fetch(`/api/budget/v2/settings?previewTarget=${target}`);
+      const res = await fetch(`/api/budget/settings?previewTarget=${target}`);
       if (res.ok) {
         const d = (await res.json()) as { data: BudgetPreviewData };
         setPreview(d.data);
@@ -308,7 +308,7 @@ function TargetDateCard({
     if (!inputValue || !/^\d{4}-\d{2}$/.test(inputValue)) return;
     setSaving(true);
     try {
-      const res = await fetch('/api/budget/v2/settings', {
+      const res = await fetch('/api/budget/settings', {
         method: 'PUT',
         headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify({ target_date: inputValue }),
@@ -425,7 +425,7 @@ export function BudgetSettingsPage({ onSettingsChange }: { onSettingsChange?: ()
       const [fixedRes, assetsRes, settingsRes] = await Promise.all([
         fetch('/api/budget/fixed-costs'),
         fetch('/api/budget/assets'),
-        fetch('/api/budget/v2/settings'),
+        fetch('/api/budget/settings'),
       ]);
       if (fixedRes.ok) {
         const d = (await fixedRes.json()) as { data: FixedCostRow[] };
